UI Improvements, Seed scoring & bug fixes

+++ balatroseeds/SeedModel.swift
@@ -14,9 +14,9 @@ final class SeedModel {
     var seed : String
     var title : String?
     var level : JokerType?
-    var score : Double?
+    var score : Int?
     
-    init(timestamp: Date, seed : String, title : String? = nil, level : JokerType? = nil, score : Double? = nil) {
+    init(timestamp: Date, seed : String, title : String? = nil, level : JokerType? = nil, score : Int? = nil) {
         self.timestamp = timestamp
         self.seed = seed
         self.title = title
